Use retention days config together with rotation strategies config
Prior to this update, if the user had retention
days configured, and the backup fell within that
window, we'd immediately retain it and move on.
However, we weren't including this backup in our
retention tracker!

Here's an example of this problem:
retention days = 7
last modified time = Feb 20th, 2025
now = Feb 24th, 2025
rotation strategy = * * * 2 * 2
The rotation strategy represents:
Keep max two backups for the month of February.

Before, in this system, we'd keep the backup made
on Feb 20th, but we would not add it to our
tracker. So what if we had other backups made on
say, Feb 5th, and Feb 10th?
All 3 would be retained, when we should only be
retaining two: the one from Feb 20th, and the one
from Feb 10th!

We should retain:
- ALL backups within the lst 7 days
- If a backup is older than 7 days, we should ONLY
  retain it if it falls within the configured
  maximums

This commit implements this logic.

It also implements more tests, to test this logic,
and improves some existing tests.